refactor: split decision-model.ts (712 lines) into 3 files under 400
  decision-model-types.ts   (115 lines) — all exported + private types
  decision-model-helpers.ts (387 lines) — all private helper functions
  decision-model.ts         (277 lines) — public API + type re-exports

No logic changed; all existing imports continue to work unchanged
because decision-model.ts re-exports everything it previously exported.
